add w3m-help-widget
