+++ packages/scaffold-html/index.ts
@@ -22,6 +22,7 @@ export * from './src/partials/w3m-connecting-wc-unsupported'
 export * from './src/partials/w3m-connecting-wc-web'
 export * from './src/partials/w3m-connecting-widget'
 export * from './src/partials/w3m-header'
+export * from './src/partials/w3m-help-widget'
 export * from './src/partials/w3m-snackbar'
 
 export { Web3ModalScaffoldHtml } from './src/client'
