+++ packages/scaffold-html/src/partials/w3m-help-widget/index.ts
@@ -1,0 +1,44 @@
+import { VisualType } from '@web3modal/ui/src/utils/TypesUtil'
+import { LitElement, html } from 'lit'
+import { customElement } from 'lit/decorators.js'
+
+type Data = {
+  images: VisualType[]
+  title: string
+  text: string
+}
+
+@customElement('w3m-help-widget')
+export class W3mHelpWidget extends LitElement {
+  // -- Members ------------------------------------------- //
+  private data: Data[] = []
+
+  // -- Render -------------------------------------------- //
+  public render() {
+    return html`
+      <wui-flex flexDirection="column" alignItems="center" gap="xl">
+        ${this.data.map(
+          item => html`
+            <wui-flex flexDirection="column" alignItems="center" gap="xl">
+              <wui-flex flexDirection="row" justifyContent="center" gap="1xs">
+                ${item.images.map(image => html`<wui-visual name=${image}></wui-visual>`)}
+              </wui-flex>
+            </wui-flex>
+            <wui-flex flexDirection="column" alignItems="center" gap="xs">
+              <wui-text variant="paragraph-500" color="fg-100" align="center">
+                ${item.title}
+              </wui-text>
+              <wui-text variant="small-500" color="fg-200" align="center">${item.text}</wui-text>
+            </wui-flex>
+          `
+        )}
+      </wui-flex>
+    `
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'w3m-help-widget': W3mHelpWidget
+  }
+}
