+++ packages/scaffold-html/src/views/w3m-what-is-a-network-view/index.ts
@@ -1,16 +1,8 @@
 import { LitElement, html } from 'lit'
 import { customElement } from 'lit/decorators.js'
-import styles from './styles'
-import { VisualType } from '@web3modal/ui/src/utils/TypesUtil'
 import { RouterController } from '@web3modal/core'
 
-type Data = {
-  images: VisualType[]
-  title: string
-  text: string
-}
-
-const data: Data[] = [
+const data = [
   {
     images: ['network', 'layers', 'system'],
     title: 'The system’s nuts and bolts',
@@ -21,14 +13,11 @@ const data: Data[] = [
     title: 'Designed for different uses',
     text: 'A wallet lets you store, send and receive digital assets like cryptocurrencies and NFTs.'
   }
-]
+] as const
 
 @customElement('w3m-what-is-a-network-view')
 export class W3mWhatIsANetworkView extends LitElement {
-  public static styles = styles
-
   // -- Render -------------------------------------------- //
-
   public render() {
     return html`
       <wui-flex
@@ -37,22 +26,7 @@ export class W3mWhatIsANetworkView extends LitElement {
         alignItems="center"
         gap="xl"
       >
-        ${data.map(
-          item =>
-            html`<wui-flex flexDirection="column" alignItems="center" gap="xl">
-                <wui-flex flexDirection="row" justifyContent="center" gap="1xs">
-                  ${item.images.map(image => html`<wui-visual name=${image}></wui-visual>`)}
-                </wui-flex>
-              </wui-flex>
-              <wui-flex flexDirection="column" alignItems="center" gap="xs">
-                <wui-text variant="paragraph-500" color="fg-100" align="center">
-                  ${item.title}</wui-text
-                >
-                <wui-text variant="small-500" color="fg-200" align="center">
-                  ${item.text}
-                </wui-text>
-              </wui-flex>`
-        )}
+        <w3m-help-widget .data=${data}></w3m-help-widget>
         <wui-button
           variant="fill"
           size="sm"
