--- packages/scaffold-html/src/views/w3m-what-is-a-network-view/styles.ts
@@ -1,7 +1,0 @@
-import { css } from 'lit'
-
-export default css`
-  wui-flex {
-    width: 100%;
-  }
-`
